update cursor hooks to capture sessions
