+++ server/internal/hooks/pending_helpers.go
@@ -154,143 +154,6 @@ func (s *Service) buildTelemetryAttributesWithMetadata(ctx context.Context, payl
 	return attrs
 }
 
-// writeCursorHookToClickHouse writes a Cursor hook event directly to ClickHouse
-// Unlike Claude hooks, Cursor payloads are already authenticated and include user_email,
-// so no Redis buffering is needed.
-func (s *Service) writeCursorHookToClickHouse(ctx context.Context, payload *gen.CursorPayload, orgID string, projectID string) {
-	attrs := s.buildCursorTelemetryAttributes(ctx, payload, orgID, projectID)
-	toolName, _ := attrs[attr.ToolNameKey].(string)
-
-	parsedProjectID, err := uuid.Parse(projectID)
-	if err != nil {
-		s.logger.ErrorContext(ctx, "Invalid project ID for Cursor hook", attr.SlogError(err))
-		return
-	}
-
-	toolInfo := telemetry.ToolInfo{
-		Name:           toolName,
-		OrganizationID: orgID,
-		ProjectID:      parsedProjectID.String(),
-		ID:             "",
-		URN:            "",
-		DeploymentID:   "",
-		FunctionID:     nil,
-	}
-
-	if s.telemetryService != nil {
-		s.telemetryService.CreateLog(telemetry.LogParams{
-			Timestamp:  time.Now(),
-			ToolInfo:   toolInfo,
-			Attributes: attrs,
-		})
-
-		s.logger.DebugContext(ctx, "Wrote Cursor hook to ClickHouse",
-			attr.SlogEvent("cursor_hook_written"),
-		)
-	}
-}
-
-// buildCursorTelemetryAttributes creates attributes for a Cursor hook event
-func (s *Service) buildCursorTelemetryAttributes(ctx context.Context, payload *gen.CursorPayload, orgID string, projectID string) map[attr.Key]any {
-	toolName := ""
-	if payload.ToolName != nil {
-		toolName = *payload.ToolName
-	}
-
-	userEmail := ""
-	if payload.UserEmail != nil {
-		userEmail = *payload.UserEmail
-	}
-
-	// Normalize to PascalCase to match Claude convention for consistent ClickHouse queries
-	hookEvent := payload.HookEventName
-	switch hookEvent {
-	case "preToolUse":
-		hookEvent = "PreToolUse"
-	case "postToolUse":
-		hookEvent = "PostToolUse"
-	case "postToolUseFailure":
-		hookEvent = "PostToolUseFailure"
-	case "beforeSubmitPrompt":
-		hookEvent = "BeforeSubmitPrompt"
-	case "afterAgentResponse":
-		hookEvent = "AfterAgentResponse"
-	case "stop":
-		hookEvent = "Stop"
-	}
-
-	attrs := map[attr.Key]any{
-		attr.EventSourceKey:    string(telemetry.EventSourceHook),
-		attr.ToolNameKey:       toolName,
-		attr.HookEventKey:      hookEvent,
-		attr.SpanIDKey:         generateSpanID(),
-		attr.TraceIDKey:        generateTraceID(),
-		attr.LogBodyKey:        fmt.Sprintf("Hook: %s", hookEvent),
-		attr.UserEmailKey:      userEmail,
-		attr.ProjectIDKey:      projectID,
-		attr.OrganizationIDKey: orgID,
-		attr.HookSourceKey:     "cursor",
-	}
-
-	if payload.Error != nil {
-		attrs[attr.HookErrorKey] = payload.Error
-	}
-
-	if payload.IsInterrupt != nil {
-		attrs[attr.HookIsInterruptKey] = *payload.IsInterrupt
-	}
-
-	// Parse MCP tool names (same mcp__ prefix convention)
-	if strings.HasPrefix(toolName, "mcp__") {
-		parts := strings.SplitN(toolName, "__", 3)
-		if len(parts) == 3 {
-			attrs[attr.ToolCallSourceKey] = parts[1]
-			attrs[attr.ToolNameKey] = parts[2]
-		}
-	}
-
-	if payload.ToolUseID != nil && *payload.ToolUseID != "" {
-		attrs[attr.TraceIDKey] = hashToolCallIDToTraceID(*payload.ToolUseID)
-	}
-	if payload.ConversationID != nil {
-		attrs[attr.GenAIConversationIDKey] = *payload.ConversationID
-	}
-	if payload.ToolUseID != nil {
-		attrs[attr.GenAIToolCallIDKey] = *payload.ToolUseID
-	}
-
-	// Store prompt text as log body for beforeSubmitPrompt events only
-	if payload.HookEventName == "beforeSubmitPrompt" && payload.Prompt != nil && *payload.Prompt != "" {
-		attrs[attr.LogBodyKey] = *payload.Prompt
-	}
-
-	// Store token usage from stop events
-	if payload.InputTokens != nil {
-		attrs[attr.GenAIUsageInputTokensKey] = *payload.InputTokens
-	}
-	if payload.OutputTokens != nil {
-		attrs[attr.GenAIUsageOutputTokensKey] = *payload.OutputTokens
-	}
-
-	// Stringify ToolInput and ToolResponse to prevent JSON path explosion in ClickHouse
-	if payload.ToolInput != nil {
-		if jsonBytes, err := json.Marshal(payload.ToolInput); err == nil {
-			attrs[attr.GenAIToolCallArgumentsKey] = string(jsonBytes)
-		} else {
-			s.logger.WarnContext(ctx, "Failed to marshal Cursor ToolInput", attr.SlogError(err))
-		}
-	}
-	if payload.ToolResponse != nil {
-		if jsonBytes, err := json.Marshal(payload.ToolResponse); err == nil {
-			attrs[attr.GenAIToolCallResultKey] = string(jsonBytes)
-		} else {
-			s.logger.WarnContext(ctx, "Failed to marshal Cursor ToolResponse", attr.SlogError(err))
-		}
-	}
-
-	return attrs
-}
-
 // flushPendingHooks retrieves all buffered hooks for a session and writes them to ClickHouse.
 // Conversation events (UserPromptSubmit, Stop) are written to PostgreSQL.
 func (s *Service) flushPendingHooks(ctx context.Context, sessionID string, metadata *SessionMetadata) {
