+++ server/internal/hooks/session_capture.go
@@ -323,256 +323,6 @@ func (s *Service) writeToolCallResultToPG(ctx context.Context, payload *gen.Clau
 	return s.insertMessageWithFallbackUpsert(ctx, metadata, chatID, projectID, msgParams, activities.DefaultClaudeChatTitle)
 }
 
-// writeCursorToolCallRequestToPG writes a Cursor tool call request (preToolUse) to PostgreSQL.
-func (s *Service) writeCursorToolCallRequestToPG(ctx context.Context, payload *gen.CursorPayload, metadata *SessionMetadata) error {
-	if payload.ConversationID == nil || *payload.ConversationID == "" {
-		return nil
-	}
-
-	projectID, err := uuid.Parse(metadata.ProjectID)
-	if err != nil {
-		return fmt.Errorf("invalid project ID: %w", err)
-	}
-
-	chatID := sessionIDToUUID(*payload.ConversationID)
-
-	toolCalls := []map[string]any{{
-		"id":   conv.PtrValOr(payload.ToolUseID, ""),
-		"type": "function",
-		"function": map[string]any{
-			"name":      conv.PtrValOr(payload.ToolName, ""),
-			"arguments": marshalToJSON(payload.ToolInput),
-		},
-	}}
-
-	toolCallsJSON, err := json.Marshal(toolCalls)
-	if err != nil {
-		return fmt.Errorf("marshal tool_calls: %w", err)
-	}
-
-	msgParams := chatRepo.CreateChatMessageParams{
-		ChatID:           chatID,
-		ProjectID:        projectID,
-		Role:             "assistant",
-		Content:          "",
-		Model:            conv.ToPGTextEmpty(conv.PtrValOr(payload.Model, "")),
-		UserID:           conv.ToPGTextEmpty(""),
-		Source:           conv.ToPGText("Cursor"),
-		ToolCalls:        toolCallsJSON,
-		FinishReason:     conv.ToPGText("tool_calls"),
-		PromptTokens:     0,
-		CompletionTokens: 0,
-		TotalTokens:      0,
-		ContentRaw:       nil,
-		ContentAssetUrl:  conv.ToPGTextEmpty(""),
-		StorageError:     conv.ToPGTextEmpty(""),
-		MessageID:        conv.ToPGTextEmpty(""),
-		ToolCallID:       conv.ToPGTextEmpty(""),
-		ExternalUserID:   conv.ToPGTextEmpty(metadata.UserEmail),
-		Origin:           conv.ToPGTextEmpty(""),
-		UserAgent:        conv.ToPGTextEmpty(""),
-		IpAddress:        conv.ToPGTextEmpty(""),
-	}
-
-	return s.insertMessageWithFallbackUpsert(ctx, metadata, chatID, projectID, msgParams, activities.DefaultCursorChatTitle)
-}
-
-// writeCursorToolCallResultToPG writes a Cursor tool call result (postToolUse/postToolUseFailure) to PostgreSQL.
-func (s *Service) writeCursorToolCallResultToPG(ctx context.Context, payload *gen.CursorPayload, metadata *SessionMetadata) error {
-	if payload.ConversationID == nil || *payload.ConversationID == "" {
-		return nil
-	}
-
-	projectID, err := uuid.Parse(metadata.ProjectID)
-	if err != nil {
-		return fmt.Errorf("invalid project ID: %w", err)
-	}
-
-	chatID := sessionIDToUUID(*payload.ConversationID)
-
-	var content string
-	if payload.HookEventName == "postToolUse" && payload.ToolResponse != nil {
-		content = marshalToJSON(payload.ToolResponse)
-	} else if payload.HookEventName == "postToolUseFailure" && payload.Error != nil {
-		content = marshalToJSON(payload.Error)
-	} else {
-		return nil
-	}
-
-	msgParams := chatRepo.CreateChatMessageParams{
-		ChatID:           chatID,
-		ProjectID:        projectID,
-		Role:             "tool",
-		Content:          content,
-		UserID:           conv.ToPGTextEmpty(""),
-		Source:           conv.ToPGText("Cursor"),
-		ToolCallID:       conv.ToPGTextEmpty(conv.PtrValOr(payload.ToolUseID, "")),
-		PromptTokens:     0,
-		CompletionTokens: 0,
-		TotalTokens:      0,
-		ContentRaw:       nil,
-		ContentAssetUrl:  conv.ToPGTextEmpty(""),
-		StorageError:     conv.ToPGTextEmpty(""),
-		Model:            conv.ToPGTextEmpty(""),
-		MessageID:        conv.ToPGTextEmpty(""),
-		ExternalUserID:   conv.ToPGTextEmpty(metadata.UserEmail),
-		FinishReason:     conv.ToPGTextEmpty(""),
-		ToolCalls:        nil,
-		Origin:           conv.ToPGTextEmpty(""),
-		UserAgent:        conv.ToPGTextEmpty(""),
-		IpAddress:        conv.ToPGTextEmpty(""),
-	}
-
-	return s.insertMessageWithFallbackUpsert(ctx, metadata, chatID, projectID, msgParams, activities.DefaultCursorChatTitle)
-}
-
-// persistCursorEventToPG routes a Cursor hook event to the appropriate PostgreSQL persistence function.
-func (s *Service) persistCursorEventToPG(ctx context.Context, payload *gen.CursorPayload, orgID string, projectID string) {
-	metadata := &SessionMetadata{
-		SessionID:   conv.PtrValOr(payload.ConversationID, ""),
-		ServiceName: "Cursor",
-		UserEmail:   conv.PtrValOr(payload.UserEmail, ""),
-		ClaudeOrgID: "",
-		GramOrgID:   orgID,
-		ProjectID:   projectID,
-	}
-
-	var err error
-	switch payload.HookEventName {
-	case "beforeSubmitPrompt":
-		err = s.persistCursorConversationEvent(ctx, payload, orgID, projectID)
-	case "afterAgentResponse":
-		err = s.persistCursorAgentResponse(ctx, payload, metadata)
-	case "preToolUse":
-		err = s.writeCursorToolCallRequestToPG(ctx, payload, metadata)
-	case "postToolUse", "postToolUseFailure":
-		err = s.writeCursorToolCallResultToPG(ctx, payload, metadata)
-	default:
-		// stop events only carry token metrics (no message content), so they are
-		// not persisted to PostgreSQL.
-		return
-	}
-
-	if err != nil {
-		s.logger.ErrorContext(ctx, "Failed to persist Cursor event to PG", attr.SlogError(err))
-	}
-}
-
-// persistCursorAgentResponse writes the assistant's response text to PostgreSQL as a chat message.
-func (s *Service) persistCursorAgentResponse(ctx context.Context, payload *gen.CursorPayload, metadata *SessionMetadata) error {
-	if payload.ConversationID == nil || *payload.ConversationID == "" {
-		return nil
-	}
-
-	content := conv.PtrValOr(payload.Text, "")
-	if content == "" {
-		return nil
-	}
-
-	projectID, err := uuid.Parse(metadata.ProjectID)
-	if err != nil {
-		return fmt.Errorf("invalid project ID: %w", err)
-	}
-
-	chatID := sessionIDToUUID(*payload.ConversationID)
-
-	msgParams := chatRepo.CreateChatMessageParams{
-		ChatID:           chatID,
-		ProjectID:        projectID,
-		Role:             "assistant",
-		Content:          content,
-		Model:            conv.ToPGTextEmpty(conv.PtrValOr(payload.Model, "")),
-		UserID:           conv.ToPGTextEmpty(""),
-		Source:           conv.ToPGText("Cursor"),
-		PromptTokens:     0,
-		CompletionTokens: 0,
-		TotalTokens:      0,
-		ContentRaw:       nil,
-		ContentAssetUrl:  conv.ToPGTextEmpty(""),
-		StorageError:     conv.ToPGTextEmpty(""),
-		MessageID:        conv.ToPGTextEmpty(""),
-		ToolCallID:       conv.ToPGTextEmpty(""),
-		ExternalUserID:   conv.ToPGTextEmpty(metadata.UserEmail),
-		FinishReason:     conv.ToPGTextEmpty(""),
-		ToolCalls:        nil,
-		Origin:           conv.ToPGTextEmpty(""),
-		UserAgent:        conv.ToPGTextEmpty(""),
-		IpAddress:        conv.ToPGTextEmpty(""),
-	}
-
-	if err := s.insertMessageWithFallbackUpsert(ctx, metadata, chatID, projectID, msgParams, activities.DefaultCursorChatTitle); err != nil {
-		return err
-	}
-
-	if s.chatTitleGenerator != nil {
-		if err := s.chatTitleGenerator.ScheduleChatTitleGeneration(
-			context.WithoutCancel(ctx),
-			chatID.String(),
-			metadata.GramOrgID,
-			metadata.ProjectID,
-		); err != nil {
-			s.logger.WarnContext(ctx, "failed to schedule chat title generation for Cursor", attr.SlogError(err))
-		}
-	}
-
-	return nil
-}
-
-// persistCursorConversationEvent writes a Cursor conversation event (user prompt) to PostgreSQL.
-func (s *Service) persistCursorConversationEvent(ctx context.Context, payload *gen.CursorPayload, orgID string, projectID string) error {
-	if payload.ConversationID == nil || *payload.ConversationID == "" {
-		s.logger.WarnContext(ctx, "Cursor conversation event missing conversation_id")
-		return nil
-	}
-
-	parsedProjectID, err := uuid.Parse(projectID)
-	if err != nil {
-		return fmt.Errorf("invalid project ID for Cursor conversation event: %w", err)
-	}
-
-	chatID := sessionIDToUUID(*payload.ConversationID)
-
-	content := conv.PtrValOr(payload.Prompt, "")
-	if content == "" {
-		return nil
-	}
-
-	metadata := &SessionMetadata{
-		SessionID:   *payload.ConversationID,
-		ServiceName: "Cursor",
-		UserEmail:   conv.PtrValOr(payload.UserEmail, ""),
-		ClaudeOrgID: "",
-		GramOrgID:   orgID,
-		ProjectID:   projectID,
-	}
-
-	msgParams := chatRepo.CreateChatMessageParams{
-		ChatID:           chatID,
-		ProjectID:        parsedProjectID,
-		Role:             "user",
-		Content:          content,
-		Model:            conv.ToPGTextEmpty(""),
-		UserID:           conv.ToPGTextEmpty(""),
-		Source:           conv.ToPGText("Cursor"),
-		PromptTokens:     0,
-		CompletionTokens: 0,
-		TotalTokens:      0,
-		ContentRaw:       nil,
-		ContentAssetUrl:  conv.ToPGTextEmpty(""),
-		StorageError:     conv.ToPGTextEmpty(""),
-		MessageID:        conv.ToPGTextEmpty(""),
-		ToolCallID:       conv.ToPGTextEmpty(""),
-		ExternalUserID:   conv.ToPGTextEmpty(metadata.UserEmail),
-		FinishReason:     conv.ToPGTextEmpty(""),
-		ToolCalls:        nil,
-		Origin:           conv.ToPGTextEmpty(""),
-		UserAgent:        conv.ToPGTextEmpty(""),
-		IpAddress:        conv.ToPGTextEmpty(""),
-	}
-
-	return s.insertMessageWithFallbackUpsert(ctx, metadata, chatID, parsedProjectID, msgParams, activities.DefaultCursorChatTitle)
-}
-
 // marshalToJSON converts any value to a JSON string.
 func marshalToJSON(v any) string {
 	if v == nil {
